More wsr keyboard grammars added and simplified an  shrunk GUI layout
Added more wsr phrases for keys like undo, redo, file save, backspace, delete, and numbered windows keys in the keyboard grammar file. Modified the GUI in the loader file to be smaller, positioned on the right side of the screen, and more compact. Implemented text wrapping and changed the layout to a single column style  by just commenting out the other column sections for now