update keyboard.py to remove misrecognition
Because i use the word moss for doing a page up key press and the word fox for the letter f, I removed the word osh, the alternate phrase for the letter o because of misrecognition with the other two words that I just mentioned. Now just saying oscar does the key press for the letter o.